Move HoverImage to hoverImage folder
